new experimental compiler pass to remove some extra calls CLJS sometimes produces
`foo.x = my.lib.create_expensive_thing(...);`

Closure might remove the name `foo.x` but the fn call will remain (if side-effectful),
sometimes the call however is useless unless you have the return value.

This adds a way to remove the call as well if the name was removed. Only works
for functions that are actually side-effecting and identified as such. Otherwise the
name removal will remove the pure parts and only leave the side-effects.